move autyhenticationProvider and authenticationManager beans to new configuration class
