+++ auth-service/src/main/java/com/chatwave/authservice/config/SecurityConfig.java
@@ -24,7 +24,6 @@ import org.springframework.security.oauth2.server.authorization.settings.ClientS
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
