+++ chat-service/src/test/java/com/chatwave/chatservice/unit/domain/MessageMapperTest.java
@@ -6,7 +6,8 @@ import com.chatwave.chatservice.domain.dto.SendMessageRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 @DisplayName("MessageMapper")
 public class MessageMapperTest {
